Move ITrigger to PlatformCore trigger namespace
