+++ Assets/PlatformCore/Infrastructure/Triggers/BaseTrigger.cs
@@ -1,15 +1,18 @@
 using System;
-using _Main.Scripts.SceneOrchestra;
 using UnityEngine;
 
-public class BaseTrigger : MonoBehaviour, ITrigger
+namespace PlatformCore.Infrastructure.Triggers
 {
-	public event Action<string> OnTriggered;
-
-	[SerializeField] private string _triggerId;
-	public string triggerId => _triggerId;
-	public void Trigger()
+	public class BaseTrigger : MonoBehaviour, ITrigger
 	{
-		OnTriggered?.Invoke(triggerId);
+		public event Action<string> OnTriggered;
+
+		[SerializeField] private string _triggerId;
+		public string triggerId => _triggerId;
+
+		public void Trigger()
+		{
+			OnTriggered?.Invoke(triggerId);
+		}
 	}
 }
